Struct rusoto_cloudsearch::CloudSearchClient

pub struct CloudSearchClient { /* fields omitted */ }

A client for the Amazon CloudSearch API.

Methods

impl CloudSearchClient

pub fn new(region: Region) -> CloudSearchClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CloudSearchClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,