Struct rusoto_cloudsearch::ExpressionStatus

pub struct ExpressionStatus {
    pub options: Expression,
    pub status: OptionStatus,
}

The value of an Expression and its current status.

Fields
options: Expression

The expression that is evaluated for sorting while processing a search request.

status: OptionStatus