Struct rusoto_cloudsearch::IndexField

pub struct IndexField {
    pub date_array_options: Option<DateArrayOptions>,
    pub date_options: Option<DateOptions>,
    pub double_array_options: Option<DoubleArrayOptions>,
    pub double_options: Option<DoubleOptions>,
    pub index_field_name: String,
    pub index_field_type: String,
    pub int_array_options: Option<IntArrayOptions>,
    pub int_options: Option<IntOptions>,
    pub lat_lon_options: Option<LatLonOptions>,
    pub literal_array_options: Option<LiteralArrayOptions>,
    pub literal_options: Option<LiteralOptions>,
    pub text_array_options: Option<TextArrayOptions>,
    pub text_options: Option<TextOptions>,
}

Configuration information for a field in the index, including its name, type, and options. The supported options depend on the IndexFieldType.

Fields
date_array_options: Option<DateArrayOptions> date_options: Option<DateOptions> double_array_options: Option<DoubleArrayOptions> double_options: Option<DoubleOptions> index_field_name: String

A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options.

Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported.

The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

index_field_type: String int_array_options: Option<IntArrayOptions> int_options: Option<IntOptions> lat_lon_options: Option<LatLonOptions> literal_array_options: Option<LiteralArrayOptions> literal_options: Option<LiteralOptions> text_array_options: Option<TextArrayOptions> text_options: Option<TextOptions>