Crate rusoto_cloudsearch

Amazon CloudSearch

If you're using the service, you're probably looking for CloudSearchClient and CloudSearch.

Structs

AccessPoliciesStatus
The configured access rules for the domain's document and search endpoints, and the current status of those rules.

AnalysisOptions
Synonyms, stopwords, and stemming options for an analysis scheme. Includes tokenization dictionary for Japanese.

AnalysisScheme
Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, JapaneseTokenizationDictionary and AlgorithmicStemming.

AnalysisSchemeStatus
The status and configuration of an AnalysisScheme.

AvailabilityOptionsStatus
The status and configuration of the domain's availability options.

BuildSuggestersRequest
Container for the parameters to the BuildSuggester operation. Specifies the name of the domain you want to update.

BuildSuggestersResponse
The result of a BuildSuggester request. Contains a list of the fields used for suggestions.

CloudSearchClient
A client for the Amazon CloudSearch API.

CreateDomainRequest
Container for the parameters to the CreateDomain operation. Specifies a name for the new search domain.

CreateDomainResponse
The result of a CreateDomainRequest. Contains the status of a newly created domain.

DateArrayOptions
Options for a field that contains an array of dates. Present if IndexFieldType specifies the field is of type date-array. All options are enabled by default.

DateOptions
Options for a date field. Dates and times are specified in UTC (Coordinated Universal Time) according to IETF RFC3339: yyyy-mm-ddT00:00:00Z. Present if IndexFieldType specifies the field is of type date. All options are enabled by default.

DefineAnalysisSchemeRequest
Container for the parameters to the DefineAnalysisScheme operation. Specifies the name of the domain you want to update and the analysis scheme configuration.

DefineAnalysisSchemeResponse
The result of a DefineAnalysisScheme request. Contains the status of the newly-configured analysis scheme.

DefineExpressionRequest
Container for the parameters to the DefineExpression operation. Specifies the name of the domain you want to update and the expression you want to configure.

DefineExpressionResponse
The result of a DefineExpression request. Contains the status of the newly-configured expression.

DefineIndexFieldRequest
Container for the parameters to the DefineIndexField operation. Specifies the name of the domain you want to update and the index field configuration.

DefineIndexFieldResponse
The result of a DefineIndexField request. Contains the status of the newly-configured index field.

DefineSuggesterRequest
Container for the parameters to the DefineSuggester operation. Specifies the name of the domain you want to update and the suggester configuration.

DefineSuggesterResponse
The result of a DefineSuggester request. Contains the status of the newly-configured suggester.

DeleteAnalysisSchemeRequest
Container for the parameters to the DeleteAnalysisScheme operation. Specifies the name of the domain you want to update and the analysis scheme you want to delete.

DeleteAnalysisSchemeResponse
The result of a DeleteAnalysisScheme request. Contains the status of the deleted analysis scheme.

DeleteDomainRequest
Container for the parameters to the DeleteDomain operation. Specifies the name of the domain you want to delete.

DeleteDomainResponse
The result of a DeleteDomain request. Contains the status of a newly deleted domain, or no status if the domain has already been completely deleted.

DeleteExpressionRequest
Container for the parameters to the DeleteExpression operation. Specifies the name of the domain you want to update and the name of the expression you want to delete.

DeleteExpressionResponse
The result of a DeleteExpression request. Specifies the expression being deleted.

DeleteIndexFieldRequest
Container for the parameters to the DeleteIndexField operation. Specifies the name of the domain you want to update and the name of the index field you want to delete.

DeleteIndexFieldResponse
The result of a DeleteIndexField request.

DeleteSuggesterRequest
Container for the parameters to the DeleteSuggester operation. Specifies the name of the domain you want to update and name of the suggester you want to delete.

DeleteSuggesterResponse
The result of a DeleteSuggester request. Contains the status of the deleted suggester.

DescribeAnalysisSchemesRequest
Container for the parameters to the DescribeAnalysisSchemes operation. Specifies the name of the domain you want to describe. To limit the response to particular analysis schemes, specify the names of the analysis schemes you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeAnalysisSchemesResponse
The result of a DescribeAnalysisSchemes request. Contains the analysis schemes configured for the domain specified in the request.

DescribeAvailabilityOptionsRequest
Container for the parameters to the DescribeAvailabilityOptions operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeAvailabilityOptionsResponse
The result of a DescribeAvailabilityOptions request. Indicates whether or not the Multi-AZ option is enabled for the domain specified in the request.

DescribeDomainsRequest
Container for the parameters to the DescribeDomains operation. By default shows the status of all domains. To restrict the response to particular domains, specify the names of the domains you want to describe.

DescribeDomainsResponse
The result of a DescribeDomains request. Contains the status of the domains specified in the request or all domains owned by the account.

DescribeExpressionsRequest
Container for the parameters to the DescribeDomains operation. Specifies the name of the domain you want to describe. To restrict the response to particular expressions, specify the names of the expressions you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeExpressionsResponse
The result of a DescribeExpressions request. Contains the expressions configured for the domain specified in the request.

DescribeIndexFieldsRequest
Container for the parameters to the DescribeIndexFields operation. Specifies the name of the domain you want to describe. To restrict the response to particular index fields, specify the names of the index fields you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeIndexFieldsResponse
The result of a DescribeIndexFields request. Contains the index fields configured for the domain specified in the request.

DescribeScalingParametersRequest
Container for the parameters to the DescribeScalingParameters operation. Specifies the name of the domain you want to describe.

DescribeScalingParametersResponse
The result of a DescribeScalingParameters request. Contains the scaling parameters configured for the domain specified in the request.

DescribeServiceAccessPoliciesRequest
Container for the parameters to the DescribeServiceAccessPolicies operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeServiceAccessPoliciesResponse
The result of a DescribeServiceAccessPolicies request.

DescribeSuggestersRequest
Container for the parameters to the DescribeSuggester operation. Specifies the name of the domain you want to describe. To restrict the response to particular suggesters, specify the names of the suggesters you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeSuggestersResponse
The result of a DescribeSuggesters request.

DocumentSuggesterOptions
Options for a search suggester.

DomainStatus
The current status of the search domain.

DoubleArrayOptions
Options for a field that contains an array of double-precision 64-bit floating point values. Present if IndexFieldType specifies the field is of type double-array. All options are enabled by default.

DoubleOptions
Options for a double-precision 64-bit floating point field. Present if IndexFieldType specifies the field is of type double. All options are enabled by default.

Expression
A named expression that can be evaluated at search time. Can be used to sort the search results, define other expressions, or return computed information in the search results.

ExpressionStatus
The value of an Expression and its current status.

IndexDocumentsRequest
Container for the parameters to the IndexDocuments operation. Specifies the name of the domain you want to re-index.

IndexDocumentsResponse
The result of an IndexDocuments request. Contains the status of the indexing operation, including the fields being indexed.

IndexField
Configuration information for a field in the index, including its name, type, and options. The supported options depend on the IndexFieldType.

IndexFieldStatus
The value of an IndexField and its current status.

IntArrayOptions
Options for a field that contains an array of 64-bit signed integers. Present if IndexFieldType specifies the field is of type int-array. All options are enabled by default.

IntOptions
Options for a 64-bit signed integer field. Present if IndexFieldType specifies the field is of type int. All options are enabled by default.

LatLonOptions
Options for a latlon field. A latlon field contains a location stored as a latitude and longitude value pair. Present if IndexFieldType specifies the field is of type latlon. All options are enabled by default.

Limits

ListDomainNamesResponse
The result of a ListDomainNames request. Contains a list of the domains owned by an account.

LiteralArrayOptions
Options for a field that contains an array of literal strings. Present if IndexFieldType specifies the field is of type literal-array. All options are enabled by default.

LiteralOptions
Options for literal field. Present if IndexFieldType specifies the field is of type literal. All options are enabled by default.

OptionStatus
The status of domain configuration option.

ScalingParameters
The desired instance type and desired number of replicas of each index partition.

ScalingParametersStatus
The status and configuration of a search domain's scaling parameters.

ServiceEndpoint
The endpoint to which service requests can be submitted.

Suggester
Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

SuggesterStatus
The value of a Suggester and its current status.

TextArrayOptions
Options for a field that contains an array of text strings. Present if IndexFieldType specifies the field is of type text-array. A text-array field is always searchable. All options are enabled by default.

TextOptions
Options for text field. Present if IndexFieldType specifies the field is of type text. A text field is always searchable. All options are enabled by default.

UpdateAvailabilityOptionsRequest
Container for the parameters to the UpdateAvailabilityOptions operation. Specifies the name of the domain you want to update and the Multi-AZ availability option.

UpdateAvailabilityOptionsResponse
The result of a UpdateAvailabilityOptions request. Contains the status of the domain's availability options.

UpdateScalingParametersRequest
Container for the parameters to the UpdateScalingParameters operation. Specifies the name of the domain you want to update and the scaling parameters you want to configure.

UpdateScalingParametersResponse
The result of a UpdateScalingParameters request. Contains the status of the newly-configured scaling parameters.

UpdateServiceAccessPoliciesRequest
Container for the parameters to the UpdateServiceAccessPolicies operation. Specifies the name of the domain you want to update and the access rules you want to configure.

UpdateServiceAccessPoliciesResponse
The result of an UpdateServiceAccessPolicies request. Contains the new access policies.

Enums

BuildSuggestersError
Errors returned by BuildSuggesters

CreateDomainError
Errors returned by CreateDomain

DefineAnalysisSchemeError
Errors returned by DefineAnalysisScheme

DefineExpressionError
Errors returned by DefineExpression

DefineIndexFieldError
Errors returned by DefineIndexField

DefineSuggesterError
Errors returned by DefineSuggester

DeleteAnalysisSchemeError
Errors returned by DeleteAnalysisScheme

DeleteDomainError
Errors returned by DeleteDomain

DeleteExpressionError
Errors returned by DeleteExpression

DeleteIndexFieldError
Errors returned by DeleteIndexField

DeleteSuggesterError
Errors returned by DeleteSuggester

DescribeAnalysisSchemesError
Errors returned by DescribeAnalysisSchemes

DescribeAvailabilityOptionsError
Errors returned by DescribeAvailabilityOptions

DescribeDomainsError
Errors returned by DescribeDomains

DescribeExpressionsError
Errors returned by DescribeExpressions

DescribeIndexFieldsError
Errors returned by DescribeIndexFields

DescribeScalingParametersError
Errors returned by DescribeScalingParameters

DescribeServiceAccessPoliciesError
Errors returned by DescribeServiceAccessPolicies

DescribeSuggestersError
Errors returned by DescribeSuggesters

IndexDocumentsError
Errors returned by IndexDocuments

ListDomainNamesError
Errors returned by ListDomainNames

UpdateAvailabilityOptionsError
Errors returned by UpdateAvailabilityOptions

UpdateScalingParametersError
Errors returned by UpdateScalingParameters

UpdateServiceAccessPoliciesError
Errors returned by UpdateServiceAccessPolicies

Traits

CloudSearch
Trait representing the capabilities of the Amazon CloudSearch API. Amazon CloudSearch clients implement this trait.

Type Definitions

APIVersion
The Amazon CloudSearch API version for a domain: 2011-02-01 or 2013-01-01.

ARN
The Amazon Resource Name (ARN) of the search domain. See Identifiers for IAM Entities in Using AWS Identity and Access Management for more information.

AlgorithmicStemming

AnalysisSchemeLanguage
An IETF RFC 4646 language code or mul for multiple languages.

AnalysisSchemeStatusList
A list of the analysis schemes configured for a domain.

Boolean

DomainId
An internally generated unique identifier for a domain.

DomainName
A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

DomainNameList
A list of domain names.

DomainNameMap
A collection of domain names.

DomainStatusList
A list that contains the status of each requested domain.

Double

DynamicFieldName

DynamicFieldNameList

ErrorCode
A machine-parsable string error or warning code.

ErrorMessage
A human-readable string error or warning message.

ExpressionStatusList
Contains the status of multiple expressions.

ExpressionValue
The expression to evaluate for sorting while processing a search request. The Expression syntax is based on JavaScript expressions. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

FieldName
A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options.

Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported.

The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

FieldNameCommaList

FieldNameList
A list of field names.

FieldValue
The value of a field attribute.

IndexFieldStatusList
Contains the status of multiple index fields.

IndexFieldType
The type of field. The valid options for a field depend on the field type. For more information about the supported field types, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

InstanceCount

Long

MaximumPartitionCount

MaximumReplicationCount

MultiAZ

OptionState

The state of processing a change to an option. One of:

RequiresIndexDocuments: The option's latest value will not be deployed until IndexDocuments has been called and indexing is complete.

Processing: The option's latest value is in the process of being activated.

Active: The option's latest value is fully deployed.

FailedToValidate: The option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

PartitionCount
The number of partitions used to hold the domain's index.

PartitionInstanceType
The instance type (such as search.m1.small) on which an index partition is hosted.

PolicyDocument
Access rules for a domain's document or search service endpoints. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide. The maximum size of a policy document is 100 KB.

SearchInstanceType
The instance type (such as search.m1.small) that is being used to process search requests.

ServiceUrl
The endpoint to which service requests can be submitted. For example, search-imdb-movies-oopcnjfn6ugofer3zx5iadxxca.eu-west-1.cloudsearch.amazonaws.com or doc-imdb-movies-oopcnjfn6ugofer3zx5iadxxca.eu-west-1.cloudsearch.amazonaws.com.

StandardName
Names must begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore).

StandardNameList

SuggesterFuzzyMatching

SuggesterStatusList
Contains the status of multiple suggesters.

UIntValue

UpdateTimestamp

Word